Refactor : swagger 응답 예시 개선, 테스트 코드 변경
