+++ build.gradle
@@ -5,7 +5,7 @@ plugins {
 }
 
 group = 'com.vsquad'
-version = '0.0.2-SNAPSHOT'
+version = '0.0.4-SNAPSHOT'
 
 java {
     sourceCompatibility = '11'
