+++ src/main/java/com/vsquad/iroas/aggregate/dto/response/ResCreatorMapDto.java
@@ -1,5 +1,6 @@
-package com.vsquad.iroas.aggregate.dto;
+package com.vsquad.iroas.aggregate.dto.response;
 
+import com.vsquad.iroas.aggregate.dto.CreatorMapDto;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
