+++ src/main/java/com/vsquad/iroas/aggregate/dto/response/ResRankingDto.java
@@ -1,5 +1,6 @@
-package com.vsquad.iroas.aggregate.dto;
+package com.vsquad.iroas.aggregate.dto.response;
 
+import com.vsquad.iroas.aggregate.dto.PlayTimeDto;
 import com.vsquad.iroas.aggregate.entity.Player;
 import com.vsquad.iroas.aggregate.entity.Ranking;
 import io.swagger.v3.oas.annotations.media.Schema;
