+++ src/main/java/com/vsquad/iroas/aggregate/dto/response/ResponseDto.java
@@ -1,4 +1,4 @@
-package com.vsquad.iroas.aggregate.dto;
+package com.vsquad.iroas.aggregate.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
@@ -10,11 +10,11 @@ import lombok.ToString;
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
-public class ResTokenDto {
+public class ResponseDto<T> {
 
-    @Schema(name = "토큰")
-    private String token;
+    @Schema(name = "body", description = "응답 몸체111")
+    private T body;
 
-    @Schema(name = "응답 메시지")
+    @Schema(name = "message", description = "응답 메시지")
     private String message;
 }
