+++ src/main/java/com/vsquad/iroas/controller/PlayerController.java
@@ -3,13 +3,17 @@ package com.vsquad.iroas.controller;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.vsquad.iroas.aggregate.dto.*;
+import com.vsquad.iroas.aggregate.dto.request.ReqPlayerDto;
+import com.vsquad.iroas.aggregate.dto.response.ResMessageDto;
+import com.vsquad.iroas.aggregate.dto.response.ResPlayerInfoDto;
+import com.vsquad.iroas.aggregate.dto.response.ResTokenDto;
+import com.vsquad.iroas.aggregate.dto.response.ResponseDto;
 import com.vsquad.iroas.config.exception.PlayerNotFoundException;
 import com.vsquad.iroas.config.exception.SteamUserNotFoundException;
 import com.vsquad.iroas.service.PlayerService;
 import com.vsquad.iroas.service.auth.CustomTokenProviderService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.Parameters;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
