+++ src/test/java/com/vsquad/iroas/controller/RankingControllerTest.java
@@ -2,7 +2,7 @@ package com.vsquad.iroas.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
-import com.vsquad.iroas.aggregate.dto.ReqRankingDto;
+import com.vsquad.iroas.aggregate.dto.request.ReqRankingDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
